chore: migrate common dict integration tests
