+++ packages/client-sdk-nodejs/test/integration/integration-setup.ts
@@ -54,6 +54,7 @@ export function SetupIntegrationTest(): {
   beforeAll(async () => {
     // Use a fresh client to avoid test interference with setup.
     const momento = momentoClientForTesting();
+    await new Promise(r => setInterval(r, 250));
     await deleteCacheIfExists(momento, cacheName);
     const createResponse = await momento.createCache(cacheName);
     if (createResponse instanceof CreateCache.Error) {
