+++ packages/client-sdk-nodejs/test/integration/shared/dictionary.test.ts
@@ -1,0 +1,6 @@
+import {runDictionaryTests} from '@gomomento/common-integration-tests';
+import {SetupIntegrationTest} from '../integration-setup';
+
+const {Momento, IntegrationTestCacheName} = SetupIntegrationTest();
+
+runDictionaryTests(Momento, IntegrationTestCacheName);
