+++ packages/common-integration-tests/src/index.ts
@@ -3,3 +3,4 @@ export * from './create-delete-list-cache';
 export * from './get-set-delete';
 export * from './generate-api-token';
 export * from './list';
+export * from './dictionary';
