+++ packages/core/src/internal/clients/cache/ICacheClient.ts
@@ -15,6 +15,13 @@ import {
   CacheListPopFront,
   CacheListRemoveValue,
   CacheListRetain,
+  CacheDictionarySetField,
+  CacheDictionarySetFields,
+  CacheDictionaryGetField,
+  CacheDictionaryFetch,
+  CacheDictionaryIncrement,
+  CacheDictionaryRemoveField,
+  CacheDictionaryRemoveFields,
 } from '../../../index';
 import {
   ScalarCallOptions,
@@ -22,6 +29,7 @@ import {
   ListRetainCallOptions,
   BackTruncatableCallOptions,
   FrontTruncatableCallOptions,
+  CollectionCallOptions,
 } from '../../../utils';
 
 // Type aliases to differentiate the different methods' optional arguments.
@@ -33,9 +41,9 @@ export type ListPushBackOptions = FrontTruncatableCallOptions;
 export type ListPushFrontOptions = BackTruncatableCallOptions;
 // export type SetAddElementOptions = CollectionCallOptions;
 // export type SetAddElementsOptions = CollectionCallOptions;
-// export type DictionarySetFieldOptions = CollectionCallOptions;
-// export type DictionarySetFieldsOptions = CollectionCallOptions;
-// export type DictionaryIncrementOptions = CollectionCallOptions;
+export type DictionarySetFieldOptions = CollectionCallOptions;
+export type DictionarySetFieldsOptions = CollectionCallOptions;
+export type DictionaryIncrementOptions = CollectionCallOptions;
 export type IncrementOptions = ScalarCallOptions;
 // export type SortedSetPutElementOptions = CollectionCallOptions;
 // export type SortedSetPutElementsOptions = CollectionCallOptions;
@@ -118,4 +126,50 @@ export interface ICacheClient extends IControlClient {
     listName: string,
     options?: ListRetainCallOptions
   ): Promise<CacheListRetain.Response>;
+  dictionarySetField(
+    cacheName: string,
+    dictionaryName: string,
+    field: string | Uint8Array,
+    value: string | Uint8Array,
+    options?: DictionarySetFieldOptions
+  ): Promise<CacheDictionarySetField.Response>;
+  dictionarySetFields(
+    cacheName: string,
+    dictionaryName: string,
+    elements:
+      | Map<string | Uint8Array, string | Uint8Array>
+      | Record<string, string | Uint8Array>,
+    options?: DictionarySetFieldsOptions
+  ): Promise<CacheDictionarySetFields.Response>;
+  dictionaryGetField(
+    cacheName: string,
+    dictionaryName: string,
+    field: string | Uint8Array
+  ): Promise<CacheDictionaryGetField.Response>;
+  dictionaryGetFields(
+    cacheName: string,
+    dictionaryName: string,
+    fields: string[] | Uint8Array[]
+  ): Promise<CacheDictionaryGetField.Response>;
+  dictionaryFetch(
+    cacheName: string,
+    dictionaryName: string
+  ): Promise<CacheDictionaryFetch.Response>;
+  dictionaryIncrement(
+    cacheName: string,
+    dictionaryName: string,
+    field: string | Uint8Array,
+    amount?: number,
+    options?: DictionaryIncrementOptions
+  ): Promise<CacheDictionaryIncrement.Response>;
+  dictionaryRemoveField(
+    cacheName: string,
+    dictionaryName: string,
+    field: string | Uint8Array
+  ): Promise<CacheDictionaryRemoveField.Response>;
+  dictionaryRemoveFields(
+    cacheName: string,
+    dictionaryName: string,
+    fields: string[] | Uint8Array[]
+  ): Promise<CacheDictionaryRemoveFields.Response>;
 }
